requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
Flask==3.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
